Refactored functions and structs into a utility module (#5788)
