+++ internal/olm/operator/bundleupgrade/upgrade.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/operator-framework/api/pkg/operators/v1alpha1"
 	registrybundle "github.com/operator-framework/operator-registry/pkg/lib/bundle"
+	fbcutil "github.com/operator-framework/operator-sdk/internal/olm/fbcutil"
 	"github.com/operator-framework/operator-sdk/internal/olm/operator"
 	"github.com/operator-framework/operator-sdk/internal/olm/operator/registry"
 	"github.com/spf13/pflag"
@@ -84,12 +85,12 @@ func (u *Upgrade) setup(ctx context.Context) error {
 	// defer defaulting the bundle add mode to after the existing CatalogSource is retrieved.
 	u.IndexImageCatalogCreator.PackageName = u.OperatorInstaller.PackageName
 	u.IndexImageCatalogCreator.BundleImage = u.BundleImage
-	u.IndexImageCatalogCreator.IndexImage = registry.DefaultIndexImage
+	u.IndexImageCatalogCreator.IndexImage = fbcutil.DefaultIndexImage
 
 	if _, hasChannelMetadata := labels[registrybundle.ChannelsLabel]; hasChannelMetadata {
 		u.IndexImageCatalogCreator.ChannelName = strings.Split(labels[registrybundle.ChannelsLabel], ",")[0]
 	} else {
-		u.IndexImageCatalogCreator.ChannelName = registry.DefaultChannel
+		u.IndexImageCatalogCreator.ChannelName = fbcutil.DefaultChannel
 	}
 
 	return nil
